requirements.txt
Django
psycopg
psycopg-binary
djangorestframework